requirements.txt
openpyxl==3.0.5
pandas==1.1.4
PyYAML==3.12
requests==2.22.0
requests-toolbelt==0.9.1
urllib3==1.22
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0